% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeJqueryUI.R
\name{includeJqueryUI}
\alias{includeJqueryUI}
\title{Inject necessary js and css assets to the head of a shiny document}
\usage{
includeJqueryUI()
}
\value{
A shiny head tag with necessary js and css assets.
}
\description{
This function has to be called within the \code{ui} of a shiny document before the
usage of other \code{shinyjqui} functions.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      includeJqueryUI(),
      # other ui codes
    ),
    server = function(input, output) {
      # server codes
    }
  )
}
}

