% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{Interactions_initializer}
\alias{Interactions_initializer}
\alias{jqui_draggabled}
\alias{jqui_droppabled}
\alias{jqui_resizabled}
\alias{jqui_selectabled}
\alias{jqui_sortabled}
\title{Initiate jQuery UI interactions}
\usage{
jqui_resizabled(tag, options = NULL)

jqui_sortabled(tag, options = NULL)

jqui_draggabled(tag, options = NULL)

jqui_droppabled(tag, options = NULL)

jqui_selectabled(tag, options = NULL)
}
\arguments{
\item{tag}{A shiny tag to enable interaction.}

\item{options}{A list of interaction options. Please see
\url{http://api.jqueryui.com/category/interactions/} for more details.}
}
\value{
A modified shiny tag with interation enabled.
}
\description{
These functions should be used in \code{ui} of a shiny document that initiate
interactions of shiny tag element(s).
}

