% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{Interactions_switcher}
\alias{Interactions_switcher}
\alias{jqui_draggable}
\alias{jqui_droppable}
\alias{jqui_resizable}
\alias{jqui_selectable}
\alias{jqui_sortable}
\title{Enable or disable jQuery UI interactions}
\usage{
jqui_sortable(selector, switch = TRUE, options = NULL)

jqui_draggable(selector, switch = TRUE, options = NULL)

jqui_droppable(selector, switch = TRUE, options = NULL)

jqui_selectable(selector, switch = TRUE, options = NULL)

jqui_resizable(selector, switch = TRUE, options = NULL)
}
\arguments{
\item{selector}{A \href{https://api.jquery.com/category/selectors/}{jQuery's
selector} that determines the shiny tag element(s) whose interaction is
enable or disable interactions.}

\item{switch}{A boolean value to determine whether to enable or disable an
interaction.}

\item{options}{A list of interaction options. Ignored when \code{switch} is
set as \code{FALSE}. Please see
\url{http://api.jqueryui.com/category/interactions/} for more details.}
}
\description{
These functions should be used in \code{server} of a shiny document that
enable or disable interactions of shiny tag element(s).
}

