% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_helper.R
\name{jqui_drag}
\alias{jqui_drag}
\title{Change the position of an element relative to another}
\usage{
jqui_drag(selector, my = "center", at = "center", of = "body",
  collision = "flip", using = NULL, within = "body")
}
\arguments{
\item{selector}{A
\href{https://api.jquery.com/category/selectors/}{jQuery_selector} that defines
the element(s) of which the position to change.}

\item{my}{A string that defines which position on the element being
positioned to align with the target element, e.g., \code{"center"}, \code{"right center"}, \code{"left bottom"}, \code{"right+10 top-25"}.}

\item{at}{A string that defines which position on the target element to align
the positioned element against, e.g., \code{"center"}, \code{"right center"}, \code{"left bottom"}, \code{"right+10 top-25"}.}

\item{of}{Defines the target element to position against. Can be either a
\href{https://api.jquery.com/category/selectors/}{jQuery_selector} or a \code{JS()}
wrapped \href{http://api.jquery.com/Types/#jQuery}{jQuery_object}}

\item{collision}{A string. When the positioned element overflows the window
in some direction, move it to an alternative position, e.g., \code{"flip"},
\code{"fit"}, \code{"fit flip"}, \code{"fit none"}.}

\item{using}{A \code{JS()} wrapped JavaScript callback function. When specified,
the actual property setting is delegated to this callback.}

\item{within}{Element to position within, affecting collision detection.}
}
\description{
Change the position of the positioned element(s) relative to target element
by JavaScript. This is a wrapper of jQuery UI function
\href{http://api.jqueryui.com/position/}{.position()}. Please refer to the
function document for more details.
}
