% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_interactions.R
\name{jqui_enable_bookmarking}
\alias{jqui_enable_bookmarking}
\title{Enable bookmarking state of mouse interactions}
\usage{
jqui_enable_bookmarking()
}
\description{
Enable \href{https://shiny.rstudio.com/articles/bookmarking-state.html}{shiny bookmarkingstate} of elements
with mouse interactions initiated. By calling this function in server, the
position, size, selection state and sorting state of elements can be saved
and restored through an URL.
}
\details{
\strong{Limitations:}
\itemize{
\item The target element should have an \code{id}
\item Cann't restore items who were moved from other sortable elements with
\code{connectWith} settings
}
}
