% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{jqui_icon}
\alias{jqui_icon}
\title{Create a jQuery UI icon}
\usage{
jqui_icon(name)
}
\arguments{
\item{name}{Class name of icon. The "ui-icon-" prefix can be omitted (i.e.
use "ui-icon-flag" or "flag" to display a flag icon)}
}
\value{
An icon element
}
\description{
Create an jQuery UI pre-defined icon. For lists of available icons, see
\url{http://api.jqueryui.com/theming/icons/}.
}
\examples{
jqui_icon('caret-1-n')

library(shiny)

# add an icon to an actionButton
actionButton('button', 'Button', icon = jqui_icon('refresh'))

# add an icon to a tabPanel
tabPanel('Help', icon = jqui_icon('help'))
}
