% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_helper.R
\name{jqui_resize}
\alias{jqui_resize}
\title{Change the size of an element}
\usage{
jqui_resize(selector, width = NULL, height = NULL)
}
\arguments{
\item{selector}{A
\href{https://api.jquery.com/category/selectors/}{jQuery_selector} that defines
the target element(s).}

\item{width}{The CSS width to set to the target element(s). Can be a either a
string, a number or a \code{JS()} wrapped JavaScript function. See
\href{http://api.jquery.com/width/#width2}{details}.}

\item{height}{The CSS height to set to the target element(s). Can be either a
string, a number or a \code{JS()} wrapped JavaScript function. See
\href{http://api.jquery.com/height/#height2}{details}.}
}
\description{
Change the width and height of target element(s) by JavaScript.
}
\details{
Somehow, in Firefox, this function is no working when the target element is a
shiny output.
}
