% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_helper.R
\name{jqui_sort}
\alias{jqui_sort}
\title{Reorder the children of target element(s)}
\usage{
jqui_sort(selector, items)
}
\arguments{
\item{selector}{A
\href{https://api.jquery.com/category/selectors/}{jQuery_selector} that defines
the element(s) whose children to sort.}

\item{items}{A integer vector that defines the order to sort. For example,
\code{c(3, 2, 1)} means to sort the first three children elements in the
reversed order.}
}
\description{
Sort the children of target element(s) in a given order.
}
