% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyjqui-package.R
\docType{package}
\name{shinyjqui-package}
\alias{shinyjqui}
\alias{shinyjqui-package}
\title{shinyjqui: 'jQuery UI' Interactions and Effects for Shiny}
\description{
An extension to shiny that brings interactions and animation effects from
    'jQuery UI' library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yang-tang/shinyjqui}
  \item \url{https://yang-tang.github.io/shinyjqui}
  \item Report bugs at \url{https://github.com/yang-tang/shinyjqui/issues}
}

}
\author{
\strong{Maintainer}: Yang Tang \email{tang_yang@outlook.com}

}
\keyword{internal}
