% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortableTableOutput.R
\name{sortableTableOutput}
\alias{sortableTableOutput}
\title{Create a table output element with sortable rows}
\usage{
sortableTableOutput(outputId)
}
\arguments{
\item{outputId}{output variable to read the table from}
}
\value{
A table output element that can be included in a panel
}
\description{
Render a standard HTML table with table rows sortable by drag and drop. The
order of table rows is recorded in \code{input$outputId_order}.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  shinyApp(
    ui = fluidPage(
      verbatimTextOutput("order"),
      sortableTableOutput("tbl")
    ),
    server = function(input, output) {
      output$order <- renderPrint({input$tbl_order})
      output$tbl <- renderTable(mtcars, rownames = TRUE)
    }
  )
}

}
\seealso{
\code{\link[=sortableRadioButtons]{sortableRadioButtons()}}, \code{\link[=sortableCheckboxGroupInput]{sortableCheckboxGroupInput()}},
\code{\link[=sortableTabsetPanel]{sortableTabsetPanel()}}, \code{\link[=selectableTableOutput]{selectableTableOutput()}}
}
