% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_effects.R
\name{get_jqui_effects}
\alias{get_jqui_effects}
\title{Get available animation effects.}
\usage{
get_jqui_effects()
}
\value{
A character vector of effect names
}
\description{
Use this function to get all animation effects in jQuery UI.
}
