% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{jqui_bookmarking}
\alias{jqui_bookmarking}
\title{Enable bookmarking state of mouse interactions}
\usage{
jqui_bookmarking()
}
\description{
Enable shiny
\href{https://shiny.rstudio.com/articles/bookmarking-state.html}{bookmarking_state}
of mouse interactions. By calling this function in \code{server}, the elements'
\code{position}, \code{size}, \code{selection state} and \code{sorting state} changed by mouse
operations can be saved and restored through an URL.
}
