% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{orderInput}
\alias{orderInput}
\title{Create a items list input control whose order can be changed.}
\usage{
orderInput(inputId, label, items, as_source = FALSE, connect = NULL,
  item_class = c("default", "primary", "success", "info", "warning",
  "danger"), placeholder = NULL, width = "500px", ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the current
order of items.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{items}{A list or vector of items to be initially displayed. If named,
the names are displayed to user.}

\item{as_source}{A boolean value to determine whether the \code{orderInput}
is set as source mode. If source mode, items in this \code{orderInput} can
only be dragged (copied) to the connected non-source \code{orderInput}(s)
defined by \code{connect} argument. If non-source mode, items in the
\code{orderInput} can be dragged (moved) within or toward other connected
non-source \code{orderInput}(s) defined by \code{connect} argument.}

\item{connect}{optional, a vector of \code{inputId}(s) of other
\code{orderInput}(s) that the current \code{orderInput} connects to. The
behavior of the conneted \code{orderInput}(s) depend on the
\code{as_source} argument.}

\item{item_class}{One of the
\href{http://www.w3schools.com/bootstrap/bootstrap_buttons.asp}{Bootstrap
Button Styles} to apply to each item.}

\item{placeholder}{A character string to show when there is no item left in
the \code{orderInput}.}

\item{width}{The width of the input, e.g. '400px', or '100\%'. Passed to
\code{\link[shiny]{validateCssUnit}}.}

\item{...}{Arguments passed to \code{shiny::tags$div} which is used to build
the container of the \code{orderInput}.}
}
\value{
A \code{orderInput} control that can be added to a UI definition.
}
\description{
Create a items list whose order can be changed by drag and drop within or
between \code{orderInput}(s). The current items order can be obtained from
\code{input$inputId}.
}

